02-TheToaster fix: make UiToast more independent
